Reverting the change in the BatchDaemon.
I had a really interesting observation with that change.

The last change that introduced shouldRunTheJob, actually slowed down
the Log sending speed by 8X.

If we stay in one single function, it can send 40k log lines in one
second, but with the new helper method, it can only send 5k log lines in
one second. This is really an interesting observation and definitely we
need to study more, but for now I'd like to keep the inline if
statement for the performance reason.