Clear batch config if the serialized data is corrupted.
If you attempt to store a closure into shared memory, it will error out
and leave corrupt data in shared memory. When the daemon tries to read
the corrupt data, it crashes, but currently does not clear the config.
Thus, the batch daemon will continue to crash until the shared memory is
manually cleared.

Instead, if we detect corrupt data, we will clear the corrupt data from
shared memory.