Change RequestWrapper to use constant (#609)
In the default Google App Engine environment, the phpversion function is disabled and requires fiddling around to get it to work (currently overriding ini config, suhosin config and whitelist env do not work as intended). This change moves to using PHP_VERSION which is a constant available in PHP since 5.2.7 and provides the same information as the implementation of phpversion here.