Debugger: Default source context and default uniquifier (#839)
* Fix default source context and default uniquifier.

The extSourceContext field should be empty when there is no
source_contexts.json file available.

The default uniquifier should be based off a SHA1 of available files to
debug (.php files). We add the file size as a proxy for detecting
changes.

* Add test for ensuring the default source context is empty

* Add test for recursively looking at source files to generate uniquifier

* Actually add the test for default uniquifier

* Fix snippet test. Source root needs to exist

* clean up prophecy arguments

* Fix default uniquifier test for windows.

The paths generated are different on windows so we should not check for
the exact value of the generated uniquifier.