Prefer emulator project when pub/sub emulator is detected
Since the emulator is always used when the *_EMULATOR_HOST environment
variable is set, there should not be any case where the ADC credentials
are needed.

Using the ADC credentials may cause confusion when they are set on the
CLI but not in the webserver, leading to different projectIds for the
same code.

Also see https://github.com/googleapis/google-cloud-php/issues/1739