Add support for signing strings with a Credentials instance. (googleapis/google-auth-library-php#221)
This change adds `Google\Auth\SignBlobInterface`, which provides a method for signing arbitrary bytes using a credentials instance.

The signing strategy depends on the credentials type.

* AppIdentityCredentials uses the AppIdentityService.
* GCECredentials uses [IAM signing](https://cloud.google.com/iam/credentials/reference/rest/v1/projects.serviceAccounts/signBlob).
* ServiceAccountCredentials and its JWT variant use local signing with the private key in the service account.

I've tested locally and in various google compute environments. an overview is below:

```
Flex:
    GCECredentials
    IAM

Standard 5.5
    AppIdentityCredentials
    AppIdentityService

Standard 7.2
    GCECredentials
    IAM

Compute Engine
    GCECredentials
    IAM
```

The reason for this change is to support work on Signed URLs in Google Cloud PHP, and to fill a gap in our library which is covered by other auth clients like Python and Node.