feat: Add support for Monolog 2.x
Monolog 2.x was released on August 30th, 2019 and requires PHP ^7.2.

The specific issue is that Monolog 2.x adds method return types which
are not supported by PHP 5.6. On the other hand, leaving the return
type declarations out causes PHP 7.x to throw exceptions.

AppEngineFlexHandler and AppEngineFlexFormatter remain the
implementation for Monolog 1.x.

AppEngineFlexV2Handler and AppEngineFlexV2Formatter are the new
implementations for Monolog 2.x.

The new classes are the same as the old one with the only difference
being added return type declarations.