fix(logging): Fix deprecation message in PsrLogger with PHP 8.1
In PHP 8.1, when implementing `Serializable` without implementing the magic
`serialize`/`__unserialize` methods, a deprecation warning will be generated.

This PR adds these methods and moves the logic as to not duplicate any code but still provide
BC for >= PHP 7.2. Tests remain unchanged and working.

The Serializable interface will be removed in PHP 9.0:
- RFC: https://wiki.php.net/rfc/phase_out_serializable
- Deprecation Warning docs: https://www.php.net/manual/en/class.serializable.php
>As of PHP 8.1.0, a class which implements Serializable without also implementing __serialize() and __unserialize() will generate a deprecation warning.