fix: fatal error in InsecureCredentialsWrapper with gax >= 1.28.1
Since gax 1.28.1, any attempt to use the InsecureCredentialsWrapper
(e.g. to connect to an emulator) throws a fatal error caused by:

`Typed property Google\ApiCore\CredentialsWrapper::$universeDomain
must not be accessed before initialization`

The issue was reported in googleapis/gax-php#537 related to use of
cloud-php-spanner, but in fact all usages of the class are broken.

This was caused by googleapis/gax-php#534, which added runtime
calls to the `checkUniverseDomain()` method that was previously
private. Unfortunately, that method assumes that the universeDomain
property has been initialised in the constructor. However
InsecureCredentialsWrapper does not call the parent constructor,
leaving all the internals of the class in an inconsistent /
undefined state.

Patching this method to a no-op fixes the immediate issue.